Benchmarks: try fresh topic for batched consumer 5000k
